test: add IocPvDisplayTest pattern with TextOutput pilot
Output components don't write to PVs, so the integration test pattern
is "write via a sibling TextInput, verify the readback display shows
the same value". Renamed TextOutput.stories.js -> .jsx so the new
render function can use JSX.

Both TextInput and TextOutput render as MUI <TextField> (role=textbox);
disambiguate the queries with distinct labels ("seed" vs "readback")
and findByLabelText. Writes 5005, asserts both display 5005, then
writes 5000 and re-asserts so we know the readback is genuinely
following PV updates rather than stuck on the first value it saw.
